test: relocate runtime-scan document search test out of sibyl-core
test_search_documents_runtime_scan_stops_at_chunk_budget (added in
PR #64) lived in the sibyl-core test suite but exercises
_search_documents_runtime_scan, which lazily imports
sibyl.persistence.content_runtime. The core package's isolated CI
environment has no sibyl module, so the test failed core:test-cov
with "ModuleNotFoundError: No module named 'sibyl'" and turned main
red. It passed pre-merge only because a full-workspace venv had the
api package installed.

Move the test to apps/api/tests, where both sibyl and sibyl_core are
importable, and drop the now-unused contextlib/types/uuid imports
from the core test module.
